requirements_enhanced.txt
# OpenCapture 核心依赖
pynput>=1.7.6
mss>=9.0.1
Pillow>=10.0.0

# macOS 特定依赖
pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'
pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'

# Qwen3-VL 集成依赖
aiohttp>=3.9.0
pyyaml>=6.0.1

# 可选依赖（用于高级功能）
# numpy>=1.24.0  # 图像处理优化
# imagehash>=4.3.1  # 相似图片检测
# redis>=5.0.0  # 结果缓存（可选）